# openapi.yaml
openapi: 3.0.3
info:
  title: Token Metadata API
  description: Welcome to the API reference overview for the [Token Metadata
    API](https://docs.hiro.so/token-metadata-api). Service that indexes metadata
    for every SIP-009, SIP-010, and SIP-013 Token in the Stacks blockchain and
    exposes it via REST API endpoints.
  version: v2.2.1
...
paths:
...
  /metadata/v1/nft/{principal}/{token_id}:
    get:
      operationId: getNftMetadata
      summary: Non-Fungible Token Metadata
      tags:
        - Tokens
      description: Retrieves metadata for a SIP-009 Non-Fungible Token
      parameters:
        - schema:
            title: Localization
            type: string
          examples:
            es-MX:
              value: es-MX
            jp:
              value: jp
          in: query
          name: locale
          required: false
          description: Metadata localization to retrieve
        - schema:
            pattern: ^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{28,41}\.[a-zA-Z]([a-zA-Z0-9]|[-_]){0,39}$
            title: Non-Fungible Token Contract Principal
            type: string
          example: SP497E7RX3233ATBS2AB9G4WTHB63X5PBSP5VGAQ.boomboxes-cycle-12
          in: path
          name: principal
          required: true
          description: SIP-009 compliant smart contract principal
        - schema:
            title: Token ID
            type: integer
          example: "35"
          in: path
          name: token_id
          required: true
          description: Token ID to retrieve
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                title: Nft Metadata Response
                type: object
                properties:
                  token_uri:
                    format: uri
                    description: URI for this token's metadata JSON
                    type: string
                    example: ipfs://ipfs/Qmf9yDYuPTrp8NRUFf8xxDd5Ud24Dx9uYMwMn8o8G2cWPW/12200.json
                  metadata:
                    title: Metadata
                    type: object
                    required:
                      - sip
                    properties:
                      sip:
                        type: integer
                        example: 16
                      name:
                        type: string
                        example: "Satoshi's Team #12200"
                      description:
                        description: Description
                        type: string
                        example: Heavy hitters, all-stars and legends of the game join forces to create
                          a collection of unique varsity jackets
                      image:
                        format: uri
                        description: Original image URL
                        type: string
                        example: ipfs://ipfs/QmZMqhh2ztwuZ3Y8PyEp2z5auyH3TCm3nnr5ZfjjgDjd5q/12199.png
                      cached_image:
                        format: uri
                        description: Cached image URL
                        type: string
                        example: https://ipfs.io/ipfs/QmZMqhh2ztwuZ3Y8PyEp2z5auyH3TCm3nnr5ZfjjgDjd5q/12199.png
                      cached_thumbnail_image:
                        format: uri
                        description: Cached image URL
                        type: string
                        example: https://ipfs.io/ipfs/QmZMqhh2ztwuZ3Y8PyEp2z5auyH3TCm3nnr5ZfjjgDjd5q/12199.png
                      attributes:
                        type: array
                        items:
                          title: Metadata Attribute
                          type: object
                          required:
                            - trait_type
                            - value
                          properties:
                            trait_type:
                              type: string
                              example: Background
                            display_type:
                              type: string
                              example: string
                            value:
                              title: Metadata Value
                              example: value
                      properties:
                        title: Metadata Properties
                        type: object
                        additionalProperties:
                          title: Metadata Value
                          examples:
                            - value
                        example:
                          collection: Foo Collection
                          total_supply: "10000"
                      localization:
                        title: Metadata Localization
                        type: object
                        required:
                          - uri
                          - default
                          - locales
                        properties:
                          uri:
                            format: uri
                            type: string
                            example: http://token.com/metadata?hl={locale}
                          default:
                            type: string
                            example: en
                          locales:
                            type: array
                            items:
                              type: string
                            example:
                              - en
                              - jp
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                title: Not Found Error Response
                anyOf:
                  - title: Token Not Found Response
                    type: object
                    required:
                      - error
                    properties:
                      error:
                        type: string
                        enum:
                          - Token not found
                  - title: Contract Not Found Response
                    type: object
                    required:
                      - error
                    properties:
                      error:
                        type: string
                        enum:
                          - Contract not found
        "422":
          description: Default Response
          content:
            application/json:
              schema:
                title: Error Response
                anyOf:
                  - title: Token Metadata Fetch In Progress Response
                    type: object
                    required:
                      - error
                    properties:
                      error:
                        type: string
                        enum:
                          - Token metadata fetch in progress
                  - title: Locale Not Found Response
                    type: object
                    required:
                      - error
                    properties:
                      error:
                        type: string
                        enum:
                          - Locale not found
                  - type: object
                    required:
                      - error
                      - message
                    properties:
                      error:
                        type: string
                        enum:
                          - Token error
                      message:
                        type: string
  /metadata/v1/sft/{principal}/{token_id}:
    get:
      operationId: getSftMetadata
      summary: Semi-Fungible Token Metadata
      tags:
        - Tokens
      description: Retrieves metadata for a SIP-013 Semi-Fungible Token
      parameters:
        - schema:
            title: Localization
            type: string
          examples:
            es-MX:
              value: es-MX
            jp:
              value: jp
          in: query
          name: locale
          required: false
          description: Metadata localization to retrieve
        - schema:
            pattern: ^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{28,41}\.[a-zA-Z]([a-zA-Z0-9]|[-_]){0,39}$
            title: Semi-Fungible Token Contract Principal
            type: string
          example: SP3K8BC0PPEVCV7NZ6QSRWPQ2JE9E5B6N3PA0KBR9.key-alex-autoalex-v1
          in: path
          name: principal
          required: true
          description: SIP-013 compliant smart contract principal
        - schema:
            title: Token ID
            type: integer
          example: "35"
          in: path
          name: token_id
          required: true
          description: Token ID to retrieve
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                title: Sft Metadata Response
                type: object
                properties:
                  token_uri:
                    format: uri
                    description: URI for this token's metadata JSON
                    type: string
                    example: ipfs://ipfs/Qmf9yDYuPTrp8NRUFf8xxDd5Ud24Dx9uYMwMn8o8G2cWPW/12200.json
                  decimals:
                    type: integer
                    example: 6
                  total_supply:
                    type: string
                    example: "250"
                  metadata:
                    title: Metadata
                    type: object
                    required:
                      - sip
                    properties:
                      sip:
                        type: integer
                        example: 16
                      name:
                        type: string
                        example: "Satoshi's Team #12200"
                      description:
                        description: Description
                        type: string
                        example: Heavy hitters, all-stars and legends of the game join forces to create
                          a collection of unique varsity jackets
                      image:
                        format: uri
                        description: Original image URL
                        type: string
                        example: ipfs://ipfs/QmZMqhh2ztwuZ3Y8PyEp2z5auyH3TCm3nnr5ZfjjgDjd5q/12199.png
                      cached_image:
                        format: uri
                        description: Cached image URL
                        type: string
                        example: https://ipfs.io/ipfs/QmZMqhh2ztwuZ3Y8PyEp2z5auyH3TCm3nnr5ZfjjgDjd5q/12199.png
                      cached_thumbnail_image:
                        format: uri
                        description: Cached image URL
                        type: string
                        example: https://ipfs.io/ipfs/QmZMqhh2ztwuZ3Y8PyEp2z5auyH3TCm3nnr5ZfjjgDjd5q/12199.png
                      attributes:
                        type: array
                        items:
                          title: Metadata Attribute
                          type: object
                          required:
                            - trait_type
                            - value
                          properties:
                            trait_type:
                              type: string
                              example: Background
                            display_type:
                              type: string
                              example: string
                            value:
                              title: Metadata Value
                              example: value
                      properties:
                        title: Metadata Properties
                        type: object
                        additionalProperties:
                          title: Metadata Value
                          examples:
                            - value
                        example:
                          collection: Foo Collection
                          total_supply: "10000"
                      localization:
                        title: Metadata Localization
                        type: object
                        required:
                          - uri
                          - default
                          - locales
                        properties:
                          uri:
                            format: uri
                            type: string
                            example: http://token.com/metadata?hl={locale}
                          default:
                            type: string
                            example: en
                          locales:
                            type: array
                            items:
                              type: string
                            example:
                              - en
                              - jp
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                title: Not Found Error Response
                anyOf:
                  - title: Token Not Found Response
                    type: object
                    required:
                      - error
                    properties:
                      error:
                        type: string
                        enum:
                          - Token not found
                  - title: Contract Not Found Response
                    type: object
                    required:
                      - error
                    properties:
                      error:
                        type: string
                        enum:
                          - Contract not found
        "422":
          description: Default Response
          content:
            application/json:
              schema:
                title: Error Response
                anyOf:
                  - title: Token Metadata Fetch In Progress Response
                    type: object
                    required:
                      - error
                    properties:
                      error:
                        type: string
                        enum:
                          - Token metadata fetch in progress
                  - title: Locale Not Found Response
                    type: object
                    required:
                      - error
                    properties:
                      error:
                        type: string
                        enum:
                          - Locale not found
                  - type: object
                    required:
                      - error
                      - message
                    properties:
                      error:
                        type: string
                        enum:
                          - Token error
                      message:
                        type: string
...
servers:
  - url: https://api.hiro.so/
    description: mainnet
  - url: https://api.testnet.hiro.so/
    description: testnet
...